Initialize FrameworkManager in the tests that require it
