+++ Framework/PythonInterface/test/python/mantid/api/AlgorithmFactoryTest.py
@@ -11,12 +11,15 @@ import testhelpers
 
 from mantid.api import AlgorithmFactory, FrameworkManagerImpl, PythonAlgorithm
 
+
 class IsAnAlgorithm(PythonAlgorithm):
     def PyInit(self):
         pass
 
+
 class NotAnAlgorithm(object):
     pass
+
 
 class AlgorithmFactoryTest(unittest.TestCase):
 
@@ -25,6 +28,16 @@ class AlgorithmFactoryTest(unittest.TestCase):
 
     def test_get_algorithm_factory_does_not_return_None(self):
         self.assertTrue(AlgorithmFactory is not None )
+
+    def test_getDescriptors(self):
+
+        descriptors = AlgorithmFactory.getDescriptors(True)
+        self.assertGreater(len(descriptors), 0)
+        d = descriptors[0]
+        self.assertTrue(hasattr(d, 'name'))
+        self.assertTrue(hasattr(d, 'alias'))
+        self.assertTrue(hasattr(d, 'category'))
+        self.assertTrue(hasattr(d, 'version'))
 
     def test_exists_returns_correct_value_for_given_args(self):
         self.assertTrue(AlgorithmFactory.exists('ConvertUnits')) #any version
