+++ Framework/PythonInterface/test/python/plugins/algorithms/WorkflowAlgorithms/sans/SANSCreateAdjustmentWorkspacesTest.py
@@ -6,7 +6,7 @@
 # SPDX - License - Identifier: GPL - 3.0 +
 from __future__ import (absolute_import, division, print_function)
 import unittest
-import mantid
+from mantid.api import FrameworkManager
 from sans.test_helper.test_director import TestDirector
 from sans.common.general_functions import create_unmanaged_algorithm
 from sans.common.constants import EMPTY_NAME
@@ -21,6 +21,10 @@ class SANSCreateAdjustmentWorkspacesTest(unittest.TestCase):
     test_wav_min = 1.
     test_wav_max = 11.
     test_wav_width = 2.
+
+    @classmethod
+    def setUpClass(cls):
+        FrameworkManager.Instance()
 
     @staticmethod
     def _get_state():
