+++ Framework/PythonInterface/test/python/plugins/algorithms/WorkflowAlgorithms/sans/SANSNormalizeToMonitorTest.py
@@ -6,7 +6,7 @@
 # SPDX - License - Identifier: GPL - 3.0 +
 from __future__ import (absolute_import, division, print_function)
 import unittest
-import mantid
+from mantid.api import FrameworkManager
 
 from sans.test_helper.test_director import TestDirector
 from sans.state.normalize_to_monitor import get_normalize_to_monitor_builder
@@ -44,6 +44,10 @@ def get_expected_for_spectrum_1_case(monitor_workspace, selected_detector):
 
 
 class SANSNormalizeToMonitorTest(unittest.TestCase):
+
+    @classmethod
+    def setUpClass(cls):
+        FrameworkManager.Instance()
 
     @staticmethod
     def _get_monitor_workspace(data=None):
