+++ Framework/PythonInterface/test/python/plugins/algorithms/WorkflowAlgorithms/sans/SANSSliceEventTest.py
@@ -6,8 +6,8 @@
 # SPDX - License - Identifier: GPL - 3.0 +
 from __future__ import (absolute_import, division, print_function)
 import unittest
-import mantid
 
+from mantid.api import FrameworkManager
 from mantid.kernel import DateAndTime
 from mantid.dataobjects import Workspace2D
 from sans.common.general_functions import (create_unmanaged_algorithm)
@@ -46,6 +46,11 @@ def provide_workspace_with_proton_charge(is_event=True):
 
 
 class SANSSliceEventTest(unittest.TestCase):
+
+    @classmethod
+    def setUpClass(cls):
+        FrameworkManager.Instance()
+
     @staticmethod
     def _provide_workspaces(is_event=True):
         workspace = provide_workspace_with_proton_charge(is_event=is_event)
