+++ scripts/test/SANS/common/log_tagger_test.py
@@ -6,12 +6,17 @@
 # SPDX - License - Identifier: GPL - 3.0 +
 from __future__ import (absolute_import, division, print_function)
 import unittest
-import mantid
-from mantid.api import AlgorithmManager
+
+from mantid.api import AlgorithmManager, FrameworkManager
 from sans.common.log_tagger import (has_tag, set_tag, get_tag, has_hash, set_hash, get_hash_value)
 
 
 class SANSLogTaggerTest(unittest.TestCase):
+
+    @classmethod
+    def setUpClass(cls):
+        FrameworkManager.Instance()
+
     @staticmethod
     def _provide_sample_workspace():
         alg = AlgorithmManager.createUnmanaged("CreateSampleWorkspace")
